+++ daemon/logger/loggerutils/multireader/multireader.go
